compilers, include: port the long->int32 fixes to the principia lineage
Port of the kencc long->int32 campaign (see changes.txt and commits
f18b103d, 64e35582, a61e8c5f, 14de1a76) to the principia-synced 8c/5c/cc:
- include/common.out.h: Ieee l/h fields int32; the linkers copy the raw
  struct bytes into the data segment so 64-bit longs corrupted every
  emitted double (caught by tests/s/variants float_386.s)
- cc.h: Node xoffset int32 (tests/c/misc/minus_one_index.c)
- cc2/pswt.c: int32 casts in the switch sort (tests/c/misc/switch_hex.c)
- 8c/5c: Adr/Var offset, sconst, hi64v/lo64v, zaddr/zname/align/maxround,
  xoffset += casts, and the matching %ld->%d listing formats

8c and 8c___ now produce byte-identical objects on all tests/c/misc
trigger files.

+++ compilers/5c/gc.h
@@ -57,7 +57,8 @@ struct	Adr
     // enum<Operand_kind> (D_NONE by default)
     char	type;
 
-    long	offset; // generic value, offset or simple constant
+    // claude: int32, not long, like in the kencc lineage (see changes.txt)
+    int32	offset; // generic value, offset or simple constant
     char	sval[NSNAME];
     double	dval;
     Ieee	ieee;
@@ -161,7 +162,7 @@ struct	Var
     char	symkind;
     //enum<type_kind>
     char	etype;
-    long	offset;
+    int32	offset;
 };
 /*e: struct [[Var]] */
