+++ compilers/5c/swt.c
@@ -504,7 +504,7 @@ void
 zname(Biobuf *b, Sym *s, int t)
 {
     char *n, bf[7];
-    ulong sig;
+    uint32 sig;
 
     n = s->name;
     /*s: [[zname()]] if generate signature for symbol s */
@@ -540,7 +540,7 @@ zname(Biobuf *b, Sym *s, int t)
 char*
 zaddr(char *bp, Adr *a, int s)
 {
-    long l;
+    int32 l;
     Ieee e;
 
     bp[0] = a->type;
@@ -606,10 +606,10 @@ zaddr(char *bp, Adr *a, int s)
 
 /*s: function [[align]](arm) */
 /// ??? -> <>
-long
-align(long i, Type *t, int op)
+int32
+align(int32 i, Type *t, int op)
 {
-    long o;
+    int32 o;
     Type *v;
     int w;
 
@@ -679,8 +679,8 @@ align(long i, Type *t, int op)
 /*e: function [[align]](arm) */
 
 /*s: function [[maxround]] */
-long
-maxround(long max, long v)
+int32
+maxround(int32 max, int32 v)
 {
     v = round(v, SZ_LONG);
     if(v > max)
